提审前的修改

+++ Qunachi/主页/餐馆信息/View/OKShopInfoSection1Row1Cell.m
@@ -95,7 +95,7 @@
     else
     {
         self.recommentLabel.hidden=NO;
-        self.recommentLabel.text=[NSString stringWithFormat:@"推荐 %ld",_model.RecommendCount];
+        self.recommentLabel.text=[NSString stringWithFormat:@"推荐 %ld",(long)_model.RecommendCount];
     }
